### 2024-09-23 #7268 @editorialbot
[PRE REVIEW]: ratesb_python: Rate Law Analysis for SBML and Antimony
**Submitting author:** @longxf21 (<a href="http://orcid.org/0009-0006-6523-8302">Longxuan Fan</a>)
**Repository:** https://github.com/sys-bio/ratesb_python
**Branch with paper.md** (empty if default branch):
**Version:** v0.2.6
