### 2024-09-23 #7268 @Kevin-Mattheus-Moerman
@longxf21 Dear author, thanks for this submission. I am the AEiC on this track and here to help process the initial steps.
The paper compilation error above seems to relate to the fact that the paper lacks the required YAML content. Please update the paper as per our guidelines and include the required YAML content.
