### 2024-10-14 #7268 @longxf21
Hi Kevin,

I apologize that I didn't get back to you. I took some time to review the
submission guidelines and followed the Docker one. I compiled it with
Docker and it works fine now.

Thanks
Longxuan

On Mon, Oct 14, 2024 at 03:18 Kevin Mattheus Moerman <
***@***.***> wrote:

> @longxf21 <https://github.com/longxf21> have you updated the paper yet?
>
