+++ paper/paper.yaml
@@ -14,11 +14,11 @@ authors:
   - name: "Herbert M. Sauro"
     affiliation: 3
 affiliations:
-  - name: "Viterbi School of Engineering, University of Southern California"
+  - name: "Viterbi School of Engineering, University of Southern California, 3650 McClintock Avenue, Los Angeles, CA 90089, United States of America"
     index: 1
-  - name: "eScience Institute, University of Washington"
+  - name: "eScience Institute, University of Washington, 3910 15th Ave NE, Seattle, WA 98195, United States of America"
     index: 2
-  - name: "Department of Bioengineering, University of Washington"
+  - name: "Department of Bioengineering, University of Washington, 3720 15th Ave NE, Seattle, WA 98195, United States of America"
     index: 3
 date: 2024-10-14
 bibliography: paper.bib
