### 2024-11-26 #7268 @longxf21
@Kevin-Mattheus-Moerman Apologies for the delay, I've checked the compiled paper pdf and the problems are fixed! Thanks!
