### 2024-12-06 #7268 @ayush9pandey
Happy to review this submission for JOSS.
