### 2024-12-08 #7268 @RMeli
Hi @davidtourigny 👋

Would you be interested in reviewing this submission for JOSS (Journal of Open Source Software)?

You can find more information about the submission at the top of this Github issue #7268.

Thank you in advance!
