### 2024-12-30 #7268 @RMeli
Thank you @dilawar! We usually aim for 6 weeks from the start of the review, but it is not too strict.
