### 2024-12-30 #7618 @RMeli
> `ratesb_python` relies heavily on approaches employed in SBMLKinetics [@Xu2023], which uses the `sympy` package to do symbolic analysis of rate laws. `ratesb_python` refines and extends these approaches to increase the accuracy of classification and to improve performance.

Can you please describe in a detailed manner, in the paper, how `ratesb_python` refines and extends the approaches used by `SBMLKinetics`? This should also be partially included in the "Statement of need" section: what `ratesb_python` brings to the table that other existing software do not?

Another thing that pops to my eyes when reading the paper is the use of error codes: could you expand why the codes are necessary, and where to find a detailed explanation of the grouped error codes?
