### 2025-01-11 #7618 @dilawar
Hi,

I've started the review. All scripts/tests that I am going to write for this review will be available here https://github.com/dilawar/joss-review-7618-ratesb-python . I am going to open issues on the project repository and link them here. It will a bit noisy but everyone involved in the review will get an update.

English is not my native tounge so my apologies in advance for all sorts of slilly embarassing mistakes I am going to make with grammar and typos.
