### 2025-01-11 #7618 @dilawar
@RMeli I want to run the "complex example" mentioned in the README.md. Please advice which 'xml' and `json` file I should link to? Feel free to upload them here if they are not in repository. A note on what output to expect for given xml/json combo will be helpful.
