### 2025-01-16 #7618 @longxf21
> ~@RMeli~ @longxf21 I want to run the "complex example" mentioned in the README.md. Please advice which 'xml' and `json` file I should link to? Feel free to upload them here if they are not in repository. A note on expected output for the given xml/json combination will be most helpful.

I will compile and push files in examples folder into the repository soon, will keep you updated!
