### 2025-01-16 #7618 @longxf21
@RMeli Hi, I've pushed the revised pdf into the repo, thanks!
