### 2025-01-18 sys-bio/ratesb_python#9 @ayush9pandey
JOSS Review: List of dependencies
A list of dependencies will enhance the documentation of the package. The documentation's installation page may be a good place to point out the required packages: https://longxf-ratesb-python.readthedocs-hosted.com/en/latest/installation.html

I understand that these are automatically handled by the pyproject file when installing the package. But advanced users/developers may require version specific information for the dependencies. So it would be great if these were added to the documentation and readme.
