### 2025-01-19 sys-bio/ratesb_python#10 @ayush9pandey
JOSS Review: Examples in the readme
A few suggestions for the example in the README. (I really like that you included these simple examples -- very helpful to learn the package)

First -- A minor typo in the readme "ratesn_python" should be "ratesb_python"

![Image](https://github.com/user-attachments/assets/a01febd5-ed3b-453f-b1fc-7c361de86b4f)

Next, I think after the simple example code is shown, you could also mention the expected output and what it means in one sentence.

For the output of the simple example, here's what I see:

```
_J0:
  Warning 1004: Flux is not decreasing as product increases.
```
Here `_J0` is perhaps an internal variable (reaction id?) and I think it should not be shown in the output, or if it has a meaning, I'd appreciate a quick explanation in the output.

Since the second example is meant to demonstrate complex features like checking specific errors or warning codes, or to look for warnings by reaction. Would it make sense to include an actual SBML model and then show these complex features so there is some meaning associated with the outputs too?
