### 2025-01-19 sys-bio/ratesb_python#11 @ayush9pandey
JOSS Review: I suggest moving changelog and releases to GitHub's built-in tools
I noticed that the readme has a long-ish list of changelog and release data. I recommend using github tags and releases: https://github.com/sys-bio/ratesb_python/tags
for this, and perhaps an automated changelog generator that can generate a changelog.md stored on the repo. In any case, having a couple of prior release and major change updates in README is ok, but you can imagine how your number of versions will keep growing so you don't want all that information on the README.
