### 2025-01-19 sys-bio/ratesb_python#12 @ayush9pandey
JOSS Review: Contributing and community guidelines
To satisfy this JOSS criteria:

Community guidelines: Are there clear guidelines for third parties wishing to 1. Contribute to the software 2. Report issues or problems with the software 3. Seek support

I recommend a dedicated contributing.md file that follows some of the best practices discussed here: https://docs.github.com/en/communities/setting-up-your-project-for-healthy-contributions/setting-guidelines-for-repository-contributors
