### 2025-01-19 #7618 @ayush9pandey
ratesb_python is a handy tool for systems and synthetic biology modeling community of researchers. It provides users an API that can be integrated into other modeling pipeline or used standalone to verify the correctness of rate laws. Many possible applications will benefit from this package. Great work from the team!

I have raised a few issues that can be easily addressed on the original repository:
https://github.com/sys-bio/ratesb_python/issues/12
https://github.com/sys-bio/ratesb_python/issues/11
https://github.com/sys-bio/ratesb_python/issues/10
https://github.com/sys-bio/ratesb_python/issues/9

Will test the functionality of the software one more time and wait for the authors to resolve these issues before I finish my checklist.
