### 2025-01-19 sys-bio/ratesb_python#10 @dilawar
Related #7 and #8
