### 2025-01-20 sys-bio/ratesb_python#14 @ayush9pandey
Do SBML local parameters work?
I am trying to use ratesb_python with an SBML model that has local parameter definitions. My model generation tool (biocrnpyler) creates models with local parameters. Even for a simple model of degradation reactions, I am unable to verify the rate laws. So I wonder if local parameters is the issue? The SBML model that I am using is available here: https://github.com/ayush9pandey/ratesb_python_review/blob/main/test_model.xml
