### 2025-02-02 sys-bio/ratesb_python#15 @dilawar
Easy way to get `result.Result` as JSON/XML?
Following program prints `result` in plain text format.

```python
from ratesb_python import check_model

result = check_model("S->P;k1*S")
print(type(result), result)
```

```text
<class 'results.Results'> _J0:
  Warning 1004: Flux is not decreasing as product increases.
```

For many developers/users of this package, especially at cli, it would also be very useful to see `result.Results` as `JSON` or `XML` or any other structured format. If you plan to add this functionality, perhaps it can also be highlighted in the documentation. I believe that a json dump will also be useful when using this package as library and not just at CLI.

Ref: https://github.com/openjournals/joss-reviews/issues/7618
