### 2025-02-02 sys-bio/ratesb_python#16 @dilawar
Broken example in docs: Complex Example at https://longxf-ratesb-python.readthedocs-hosted.com/en/latest/quickstart.html#complex-example
https://longxf-ratesb-python.readthedocs-hosted.com/en/latest/quickstart.html#complex-example  has the following code. It is currently broken.

```python
from ratesb_python import Analyzer

analyzer = Analyzer("S1->P1; k1 * S1")

# Analyze the model for rate law correctness
analyzer.check_all()

# Display all errors and warnings
print(analyzer.results)

# Check selected errors and warnings
analyzer.checks([1, 2, 1001, 1002])

print(analyzer.results)

# Display only warnings
warnings = analyzer.results.get_all_warnings()
for reaction, messages in warnings.items():
    print(reaction, messages)

# Retrieve messages for a specific reaction
messages = results.get_messages_by_reaction("Reaction1")
print(messages)

# Remove messages for a specific reaction
results.remove_messages_by_reaction("Reaction1")

# Get number of errors and warnings
print("Num Errors: ", results.count_errors())
print("Num Warnings: ", results.count_warnings())
```

Following diff will fixed it for me (fixed file https://github.com/dilawar/joss-review-7618-ratesb-python/blob/main/review_ratesb_python/sanity_check_2.py)

```
@@ -10,11 +10,11 @@ print(analyzer.results)

 # Check selected errors and warnings
 analyzer.checks([1, 2, 1001, 1002])
-
-print(analyzer.results)
+results = analyzer.results
+print(results)

 # Display only warnings
-warnings = analyzer.results.get_all_warnings()
+warnings = results.get_warnings()
 for reaction, messages in warnings.items():
     print(reaction, messages)
```

In rst docs, prefer using `.. literalinclude:` directive (https://www.sphinx-doc.org/en/master/usage/restructuredtext/directives.html#directive-literalinclude) to show code directly from file -- maintains single source of truth. This avoid copy-paste and ensures that copies do not go out of sync.

JOSS Review: https://github.com/openjournals/joss-reviews/issues/7618
