### 2025-02-02 #7618 @dilawar
I've opened a couple of issues on the package repository and have linked them in https://github.com/openjournals/joss-reviews/issues/7618#issuecomment-2585073835.

Overall, the package is in good shape. There are decent number of tests which can always be extended to cover complex cases. Documentation is decent but formatting can be improved a bit. Especially, instead of copy-pasting code from examples, prefer refer the example/test file directly.

Following are a few nitpicks on the code/projects that I think will help improving codebase. These are optional for this review.

- Add typehints along with adding `mypy` or `ruff` to development tooling would greatly help in finding bugs and make codebase more robust.
- Prefer `pathlib` over `os.path`.
- In many internal files, authors have manually added script's parent directory path to system paths e.g.

```python
current_dir = os.path.abspath(os.path.dirname(__file__))
parent_dir = os.path.dirname(current_dir)
sys.path.append(current_dir)
sys.path.append(parent_dir)

from common import util
```

Python import system is not without its problem.  Ideally, one would expect a single line `from ratesb_python.common import util` here. Perhaps this https://setuptools.pypa.io/en/latest/userguide/development_mode.html during development will help.

I can happily recommend [uv](https://docs.astral.sh/uv/) to manage this project.

- Perhaps `black` or `ruff` can be use to automatically format the code and docstrings.
