### 2025-02-02 sys-bio/ratesb_python#9 @RMeli
https://github.com/openjournals/joss-reviews/issues/7618#issuecomment-2629296302
