+++ docs/source/installation.rst
@@ -6,3 +6,13 @@ To install `ratesb_python`, execute the following command in your terminal:
 .. code-block:: bash
 
    pip install ratesb_python
+
+Dependencies
+------------
+
+The following dependencies are required and will be installed automatically. However, if you need version-specific details, refer to the `pyproject.toml` file or use `pip show <package>` after installation.
+
+- `antimony`
+- `sbmlkinetics`
+- `sympy`
+- `python-libsbml`
