+++ .github/workflows/update-changelog.yml
@@ -1,0 +1,35 @@
+name: Update Changelog
+
+on:
+  push:
+    branches:
+      - main
+  workflow_dispatch:
+
+jobs:
+  update-changelog:
+    runs-on: ubuntu-latest
+
+    steps:
+      - name: Checkout repository
+        uses: actions/checkout@v3
+        with:
+          fetch-depth: 0  # Fetch full history for changelog generation
+
+      - name: Install dependencies (for git-cliff)
+        run: |
+          sudo apt update
+          sudo apt install -y cargo
+          cargo install git-cliff
+
+      - name: Generate Changelog
+        run: |
+          git-cliff -o CHANGELOG.md
+
+      - name: Commit and Push Changes
+        run: |
+          git config --global user.name "github-actions[bot]"
+          git config --global user.email "github-actions[bot]@users.noreply.github.com"
+          git add CHANGELOG.md
+          git commit -m "chore: update changelog [skip ci]" || echo "No changes to commit"
+          git push origin main || echo "No changes to push"
