+++ examples/simple_example.py
@@ -1,0 +1,2 @@
+from ratesb_python import check_model
+print(check_model("S->P;k1*S"))
