+++ docs/source/custom_classifications.rst
@@ -10,25 +10,4 @@ Custom Rate Law Classifications
 
 Here's an example of how to define your rate laws in the JSON file:
 
-.. code-block:: json
-
-   [
-       {
-           "name": "Unidirectional Mass Action with two reactants",
-           "expression": "compartment * parameter * reactant1 * reactant2**2",
-           "optional_symbols": ["compartment", "parameter"],
-           "power_limited_species": ["reactant1,reactant2"]
-       },
-       {
-           "name": "Michaelis Menten",
-           "expression": "compartment * parameter * reactant1 / (reactant1 + parameter)",
-           "optional_symbols": ["compartment"],
-           "power_limited_species": ["reactant1"]
-       },
-       {
-           "name": "Your own rate law",
-           "expression": "use: compartment, parameter, reactant1, reactant2, reactant3, product1, product2, product3, enzyme. Do NOT use: species",
-           "optional_symbols": ["symbols that do not have to include in a rate law"],
-           "power_limited_species": ["RateSB will pay attention to the power of these species (power matters in this case) symbols that have to be set to a certain power in a rate law (please specify power in the expression)"]
-       }
-   ]
+.. literalinclude:: /examples/custom_classifications.json
