+++ docs/source/quickstart.rst
@@ -26,53 +26,13 @@ Complex Example
 
 This example shows how to use the `Analyzer` class to check a model for rate law correctness. The `Analyzer` class provides more control over the analysis process and allows for more detailed analysis of the model.
 
-.. code-block:: python
-
-   from ratesb_python import Analyzer
-
-   analyzer = Analyzer("S1->P1; k1 * S1")
-
-   # Analyze the model for rate law correctness
-   analyzer.check_all()
-
-   # Display all errors and warnings
-   print(analyzer.results)
-
-   # Check selected errors and warnings
-   analyzer.checks([1, 2, 1001, 1002])
-
-   print(analyzer.results)
-
-   # Display only warnings
-   warnings = analyzer.results.get_all_warnings()
-   for reaction, messages in warnings.items():
-       print(reaction, messages)
-
-   # Retrieve messages for a specific reaction
-   messages = results.get_messages_by_reaction("Reaction1")
-   print(messages)
-
-   # Remove messages for a specific reaction
-   results.remove_messages_by_reaction("Reaction1")
-
-   # Get number of errors and warnings
-   print("Num Errors: ", results.count_errors())
-   print("Num Warnings: ", results.count_warnings())
+.. literalinclude:: /examples/complex_example.py
 
 This example shows how to use the custom rate law checker to check a model for rate law correctness. The custom rate law checker allows for the creation of custom rate law checks. For detailed information on creating custom rate law checks, see the Custom Rate Law Checks section.
 
 Here is the custom rate law file in JSON format:
 
-.. code-block:: json
-
-    [
-        {
-            "name": "Your rate law check (Here is Uni-Directional Mass Action, which is included)",
-            "expression": "compartment * parameter * reactant1",
-            "optional_symbols": ["compartment", "parameter"],
-            "power_limited_species": []
-        }
-    ]
+.. literalinclude:: /examples/classification_tutorial.json
 
 Here is the Python code to use the custom rate law checker:
 
