### 2025-02-18 #7618 @longxf21
I've fixed the issues regarding the docs. I will take some time look into issue [15](https://github.com/sys-bio/ratesb_python/issues/15), [14](https://github.com/sys-bio/ratesb_python/issues/14), and [11](https://github.com/sys-bio/ratesb_python/issues/11).
