### 2025-03-06 #7618 @RMeli
Hi all, I haven't heard back from any of you in a while. Can you please let me know in a couple of sentences how this review is going?

I see @dilawar completed the checklist, but there are still some outstanding items in https://github.com/openjournals/joss-reviews/issues/7618#issuecomment-2585073835.

In @ayush9pandey there are still some missing items.

@longxf21 can you please have a look at the open issues from the reviewers?

Thank you everyone!
