### 2025-03-06 #7618 @ayush9pandey
Thanks for checking in @RMeli , I am still waiting for a couple of issues to be resolved, primarily https://github.com/sys-bio/ratesb_python/issues/14
A few of the other issues I raised were closed without comment. I am looking for if they were addressed some place or whether the authors decided on better approaches that made the issue obsolete.
