### 2025-03-10 sys-bio/ratesb_python#15 @longxf21
Adding support for structured output formats like JSON/XML would greatly improve interoperability and usability of the results.Results class. I will discuss with the team and make this in the next update of the software
