### 2025-03-10 sys-bio/ratesb_python#14 @longxf21
Hi Ayush. I couldn't replicate the issue. I ran the software with this model and got:
`r0:
  Warning 1020: We recommend that these parameters start with 'k': r0_k_forward
r1:
  Warning 1020: We recommend that these parameters start with 'k': r1_k_forward
r2:
  Warning 1020: We recommend that these parameters start with 'k': r2_k_forward
`
Which looks fine to me.
