### 2025-03-10 7581abd
Suppressed python3.12 string warnings
