### 2025-04-11 sys-bio/ratesb_python#14 @ayush9pandey
Would you share your code snippet?

Here's what I have:

```
from ratesb_python import Analyzer
analyzer = Analyzer("test_model.xml")
analyzer.check_all()
results = analyzer.results

# Display all errors and warnings
print(results)
```

The output is:
```
r0:
  Warning 1002: Unrecognized rate law from the standard list.
r1:
  Warning 1002: Unrecognized rate law from the standard list.
r2:
  Warning 1002: Unrecognized rate law from the standard list.
```
