### 2025-04-11 #7618 @ayush9pandey
Hi @RMeli

I looked at the status of my review. Some of the issues that I raised earlier were closed by the authors without comment, so I was not sure if they were addressed or not. On manually checking, I think a few of them were addressed but maybe a couple were deemed not important to fix by the team, which is totally fine! I believe it is usually good practice to comment on an issue the reason for closing them.

A functionality related issue is still open (https://github.com/sys-bio/ratesb_python/issues/14), where it's perhaps on my end that something is off. I'll wait for the author to confirm on this issue, so that I can fix it and run the package successfully with the test model that I created.

A performance related comment in my main review above is still unaddressed:

In the paper, the authors have mentioned:

```
This approach provides polynomial-time complexity compared to the generally non-polynomial complexity of purely symbolic methods.
```

So, in my review above, I had mentioned:

> For the JOSS checklist on "performance claims": the paper talks about polynomial time complexity, as compared to purely symbolic approaches that have been used for rate law validation before. Please provide a comparison of the time complexity for an example system (that is large enough to show the differences in the time complexity) with `ratesb_python` and symbolic methods.

Maybe I missed a message in this long thread or the repo some place. I'd appreciate any help. I apologize for being slow on this review.
