+++ setup.cfg
@@ -1,6 +1,6 @@
 [metadata]
 name = ratesb_python
-version = 0.2.5
+version = 0.2.7
 author = Longxuan Fan
 author_email = longxuan@usc.edu
 description = package for analyzing the rate laws in an SBML or Antimony model
