+++ pyproject.toml
@@ -1,6 +1,6 @@
 [project]
 name = "ratesb_python"
-version = "0.2.5"
+version = "0.2.7"
 description = "rate law analysis for systems biology"
 readme = "README.md"
 license = {file = "LICENSE"}
