### 2025-04-13 #7618 @longxf21
Hi @ayush9pandey ,
I have made a new release on pypi addressing [issue14](https://github.com/sys-bio/ratesb_python/issues/14).
