### 2025-05-13 #7618 @RMeli
> Please provide a comparison of the time complexity for an example system (that is large enough to show the differences in the time complexity) with ratesb_python and symbolic methods.

@longxf21 I agree with the reviewer that this would be great to have, and it would be a good addition to the paper, so that readers can immediately see the advantages of your approach.
