### 2025-05-13 #7618 @ayush9pandey
Thanks for the new release of the software @longxf21 ! That fixed the issue I had earlier with the functional testing of the package.

@RMeli The only remaining checklist item in my list now is regarding the performance claim that you mentioned in your comment above. I could check that off if the authors instead preferred to tone down that claim in the paper.
