### 2025-05-13 sys-bio/ratesb_python#17 @ayush9pandey
Readme example needs an update?
In the readme, the examples follow the previous version and should be updated:

```
# Display only warnings
warnings = results.get_warnings()
for reaction, messages in warnings.items():
    print(reaction, messages)
```

The new release uses the function `get_all_warnings` I believe.
