### 2025-05-13 #7618 @ayush9pandey
Due to the changes in the new release, I created another minor issue https://github.com/sys-bio/ratesb_python/issues/17
should be easy to fix in the readme.
