### 2025-06-01 sys-bio/ratesb_python#17 @ayush9pandey
What's the new release version? I still see 0.2.7 as the latest release. Or is it fixed in the code on github but not yet released?
