### 2025-06-10 #7618 @ayush9pandey
> > Please provide a comparison of the time complexity for an example system (that is large enough to show the differences in the time complexity) with ratesb_python and symbolic methods.
>
> [@longxf21](https://github.com/longxf21) I agree with the reviewer that this would be great to have, and it would be a good addition to the paper, so that readers can immediately see the advantages of your approach.

@RMeli  I'm not sure if this was addressed but I'm willing to complete my checklist if you think that's appropriate.
