+++ paper/paper.md
@@ -35,7 +35,11 @@ Mechanistic models in systems biology are essential tools for simulating and und
 
 The  `ratesb_python` package evaluates rate laws against a library of predefined types to identify anomalies that may compromise the accuracy of mechanistic models. This process involves categorizing rate laws based on their mathematical characteristics and examining their performance within the context of the model. Such analysis enables `ratesb_python` to identify potential errors and warnings, including discrepancies in reactant usage or abnormal reaction fluxes. For example, if the rate law provided for the reaction $2 H_2 + O_2 \rightarrow 2 H_2O$ is $k_1 [H_2]^2 [O_2] [H_2O]^2$ (where $k_1$ is a constant), ``ratesb_python`` reports an error since there is no defined classification for the rate law since the products are included ($[H_2O]$).
 
-Moreover, `ratesb_python` extends beyond symbolic comparisons employed by existing tools (such as SBMLKinetics [@Xu2023]) by implementing a **randomized polynomial identity test** for identifying and classifying rate laws. This approach provides polynomial-time complexity compared to the generally non-polynomial complexity of purely symbolic methods. The added benefit is that the algorithm is not restricted to a set of pre-coded equations, thereby enabling more flexible customization: users can readily define their own rate laws and instruct `ratesb_python` to check whether a provided rate law matches a generalized or custom-defined functional form. This functionality is especially beneficial for researchers who need to handle specialized or novel rate laws that do not necessarily fit the standard templates.
+Moreover, `ratesb_python` extends beyond symbolic comparisons employed by existing tools (such as SBMLKinetics [@Xu2023]) by providing a **modular and extensible classification framework** that supports both predefined and user-defined rate laws. While symbolic methods rely on hardcoded heuristics and expensive algebraic simplification—often with high time complexity—ratesb_python introduces a structured approach that separates normalization (e.g., power-lowering, constant removal) from classification, enabling faster and more reliable matching. More importantly, the algorithm is not restricted to a set of pre-coded equations, thereby enabling more flexible customization: users can readily define their own rate laws and instruct `ratesb_python` to check whether a provided rate law matches a generalized or custom-defined functional form. This functionality is especially beneficial for researchers who need to handle specialized or novel rate laws that do not necessarily fit the standard templates.
+
+In addition to its extensibility, `ratesb_python` provides practical performance improvements. While symbolic tools like `SBMLKinetics` require nested loops over all parameter combinations (up to $O(num_param^num_param*S(n))$ symbolic comparisons, where $S(n)$ is a single comparison cost) and repeated use of sympy.simplify(), `ratesb_python` reduces this to a more controlled pipeline with factorial complexity in reactant/product permutations and exponential complexity only in the number of optional symbols. In theory, worst-case time complexity of ratesb_python is bounded by $O((num_rtct!*num_prod!)*2^k*S(n))$ (where k is the number of optional symbols). In practice, however, these bounds remain modest: most biological reactions involve no more than one or two reactants and products, and the number of optional symbols is limited to three (compartment, parameter, and enzyme), keeping the actual computational cost low even for complex rate laws.
+
[…]
