+++ paper/paper.md
@@ -37,9 +37,9 @@ The  `ratesb_python` package evaluates rate laws against a library of predefined
 
 Moreover, `ratesb_python` extends beyond symbolic comparisons employed by existing tools (such as SBMLKinetics [@Xu2023]) by providing a **modular and extensible classification framework** that supports both predefined and user-defined rate laws. While symbolic methods rely on hardcoded heuristics and expensive algebraic simplification—often with high time complexity—ratesb_python introduces a structured approach that separates normalization (e.g., power-lowering, constant removal) from classification, enabling faster and more reliable matching. More importantly, the algorithm is not restricted to a set of pre-coded equations, thereby enabling more flexible customization: users can readily define their own rate laws and instruct `ratesb_python` to check whether a provided rate law matches a generalized or custom-defined functional form. This functionality is especially beneficial for researchers who need to handle specialized or novel rate laws that do not necessarily fit the standard templates.
 
-In addition to its extensibility, `ratesb_python` provides practical performance improvements. While symbolic tools like `SBMLKinetics` require nested loops over all parameter combinations (up to $O(num_param^num_param*S(n))$ symbolic comparisons, where $S(n)$ is a single comparison cost) and repeated use of sympy.simplify(), `ratesb_python` reduces this to a more controlled pipeline with factorial complexity in reactant/product permutations and exponential complexity only in the number of optional symbols. In theory, worst-case time complexity of ratesb_python is bounded by $O((num_rtct!*num_prod!)*2^k*S(n))$ (where k is the number of optional symbols). In practice, however, these bounds remain modest: most biological reactions involve no more than one or two reactants and products, and the number of optional symbols is limited to three (compartment, parameter, and enzyme), keeping the actual computational cost low even for complex rate laws.
+In addition to its extensibility, `ratesb_python` provides practical performance improvements. While symbolic tools like `SBMLKinetics` require nested loops over all parameter combinations (up to $O(num\_param^num\_param*S(n))$ symbolic comparisons, where $S(n)$ is a single comparison cost) and repeated use of sympy.simplify(), `ratesb_python` reduces this to a more controlled pipeline with factorial complexity in reactant/product permutations and exponential complexity only in the number of optional symbols. In theory, worst-case time complexity of ratesb_python is bounded by $O((num\_rtct!*num\_prod!)*2^k*S(n))$ (where k is the number of optional symbols). In practice, however, these bounds remain modest: most biological reactions involve no more than one or two reactants and products, and the number of optional symbols is limited to three (compartment, parameter, and enzyme), keeping the actual computational cost low even for complex rate laws.
 
-In simple scenearios-such as a Michaelis-Menten rate law $M->; cell * M * V / (K + M)$—with one reactant and no products, both `ratesb_python` and symbolic methods exhibit similar performance. However, for more complex rate laws, such as a reversible Michaelis-Menten formulation. However, in more complex examples like a reversible michaelis menten rate law: $S->P; cell * (Vmax / Km1) * (S - P / Keq) / (1 + S / Km1 + P / Km2)$, the difference becomes substantial. A symbolic may would require on the order of $O(25*S(n))$ comparisons due to nested parameter loop evaluations, and each comparison becomes costlier due to increased symbolic expression complexity. In contrast, ratesb_python evaluates this example with only $O((1!*1!)*2^1*S(n))=O(2*S(n))$ operations, illustrating its high scalability and efficiency in handling real-world, biologically meaningful models.
[…]
