### 2025-06-23 #7618 @RMeli
Thanks @ayush9pandey for you review and final comments! @longxf21 I'll have a final look at the paper, but please start considering the final comments from @ayush9pandey. I agree the the multiplication sign `*` should be removed (and indeed is already not used in many equations). I also agree that some equations are quite long, and might be better off as full equations instead of being inlined. Please also make sure to define all the quantities you use in the equation, since it does not seem to be the case.
