### 2025-07-30 #7618 @RMeli
@longxf21, thanks a lot for providing the additional benchmarks, it's great to see them!

In light of them, I'd honestly remove the performance claims all together, because I don't think they are substantiated. Yes, there is one case where `ratesb` is 5.6x faster, but there are other cases where `ratesb` is 10x or 12x slower.

However, the stability is a good point. To really drive it home, I'd maybe color code the correct and wrong classifications or put in bold the correct ones, so that is clear which method is best.

@ayush9pandey @dilawar could you please chime in on the latest benchmarks in the "Experimental Comparison" section? Thanks.
