### 2025-08-02 #7618 @dilawar
@RMeli I concur. The data is not sufficient to make a solid performance claim.  The newly added table talks about complexity vs perf/stability, but complexity is not quantified in the table, e.g. number of reactions/variables etc.

IMO, table is perhaps not a great way to visualize/present such claim here. To be convinced, I'd love to see a plot (with more data) where the x-axis represents the complexity of the model, y-axis performance and/or stability.
