### 2025-08-02 #7618 @ayush9pandey
Thanks @longxf21  for adding the performance comparisons.

I agree that it is perhaps better to focus on the stability improvements of your package by highlighting the wrong classifications/bold the correct ones in prior approaches.

> However, the stability is a good point. To really drive it home, I'd maybe color code the correct and wrong classifications or put in bold the correct ones, so that is clear which method is best.

Visualization suggested by @dilawar could also make sense.
