### 2025-09-08 #7618 @longxf21
> [@RMeli](https://github.com/RMeli) I concur. The data is not sufficient to make a solid performance claim. The newly added table talks about complexity vs perf/stability, but complexity is not quantified in the table, e.g. number of reactions/variables etc.
>
> IMO, table is perhaps not a great way to visualize/present such claim here. To be convinced, I'd love to see a plot (with more data) where the x-axis represents the complexity of the model, y-axis performance and/or stability.

I removed the performance claim and only kept the stability claim. I agree that a plot would be a great to visualize the performance claim as it can be quantified. But neither complexity nor stability is easy to quantify. IMO to strengthen the claim, we could add more models that shows ratesb_python being more stable.
