### 2025-09-22 #7618 @longxf21
Hi @dilawar @ayush9pandey @RMeli ,
Thanks for the feedback and patience, I've addressed the comments as follows:
- Removed all performance claims and refocused the paper on ratesb_python's classification stability advantage.
- Updated experimental section includes a comprehensive benchmark with 13 models (28 reactions) showing ratesb_python achieves 92.86% accuracy vs SBMLKinetics' 55.36%, particularly excelling on complex multi-reaction biological systems where SBMLKinetics defaults to generic classifications.
- All benchmark code and models are available in the repository, simply follow benchmark/README.md.
