### 2025-09-22 #7618 @RMeli
Thanks @longxf21, LGTM!
